Added IFileDialog and more buttons
The IFileDialog was a lot harder than anticipated. I have done a little COM programming before in C with well documented APIs but working out how to do this in rust was just a little more complicated than I expected. Compounded by things like not realising when utf16s were being returned etc., but it is up and running now.